rdma-core-44.0:

Updates from version 43.0
   * tests: Fix assert syntax in crypto tests
   * tests: Add dr action flow sampler test
   * pyverbs: Add dr-action flow sampler support
   * tests: Generalize create and query counter functions
   * tests: Add huge page allocation tests
   * tests: Don't use empty lists as default param value
   * pyverbs: Expose UAR C pointer
   * libhns: Fix multiple assignment of WC status
   * libhns: Fix return value when creating unsupported CQ flags
   * libhns: Fix immediately error sign type in data path.
   * libhns: Fix return value when creating unsupported QP types
   * debian: Add 64-bit MIPS architectures to COHERENT_DMA_ARCHS
   * mlx5: DR, Fix partial definer2 ip_version match
   * mlx5: DR, Fix error flow in dr_arg_mngr_create
   * mlx5: DR, Removed unused member of args
   * verbs: Make integer conversions explicit
   * tests: Add a python test for atomic write
   * pyverbs: Support atomic write
   * providers/rxe: Introduce ibv_wr_atomic_write()
   * tests: Fix failure in importing scapy module
   * util: mmio: fix build on MIPS with binutils >= 2.35
   * verbs: Add atomic write flags
   * Update kernel headers
   * tests: Add a test of RDMA zero-byte Write
   * tests: Add RDMA READ zero length scenario
   * util: refactor bitmap_find_first_bit()
   * util: fix invalid memory access in bitmap functions
   * util: add bitmap tests
   * mana: Microsoft Azure Network Adapter (MANA) RDMA provider
   * Update kernel headers
   * azp: Update to Ubuntu Jammy
   * azp: Use the python 3.10 format for pkg_resources.parse_version
   * sparse: Update diffs to support glibc 35
   * sparse: Fixup gen-sparse.py to generate the diffs properly
   * cbuild: Add Ubuntu 'jammy' 22.04 LTS
   * cbuild: Add centos9 stream
   * cbuild: Update for centos8 to use stream
   * cbuild: Update to Fedora 37
   * debian: Update compat to 10
   * build: Detect non-working compilers early
   * build: Remove -Wno-stringop-overread
   * verbs: Pass the cmd to execute_cmd through a void *
   * util: Enable uninitialized_var() for newer gcc too
   * ocrdma: Remove unused discard_cnt
   * irdma: Remove unused _total_size
   * acm: acm_msg::ep_data is not an array
   * verbs: Remove double __attribute_const in mbps_to_ibv_rate()
   * sparse: Deal with the new sparse warning "array of flexible structures"
   * Fix incorrect string lengths in sscanf
   * Add clangd stuff to .gitignore
   * azp: Tell checkpatch where the git dir is
   * tests: Add test for multi blocks AES-XTS encryption
   * pyverbs: Add multi block AES-XTS enum
   * mlx5: Add support for multi-block AES-XTS encryption
   * mlx5: DR, Add memory statistics for domain object
   * mlx5: DR, Calculate the threshold of each pool according to its type
   * mlx5: DR, Free vport ib_ports on domain caps uninitialization
   * mlx5: DR, Fix argument put list access
   * mlx5: DR, Clear args send flags
   * mlx5: DR, Fix a memory leak upon destroying the arguments' pools
   * mlx5: DR, Enable up to 64 modifications in modify-header action
   * mlx5: DR, Enable pattern and arguments for devices above ConnectX-6 DX
   * mlx5: DR, Fix issue in writing pattern for accelerate modify list
   * mlx5: DR, Apply modify action and decapl3 correctly
   * mlx5: DR, Initialize pd_num before pattern/args initialization
   * mlx5: DR, Use the same STE for cvlan and svlan match
   * providers:rxe: Fix -Errors
   * mlx5: Fix bugs when registering a MR in vfio context
   * mlx5: Allocate one more pointer in mlx5dv_get_vfio_device_list
   * mlx5: Fix memory corruption at writing UMR WQE
   * ibdiags: Fix max length of node description (dump_fst/ibroute/ibtracert)
   * providers/rxe: Implement create_srq_ex
   * redhat: Support rpmbuild on RHEL9
   * documentation: Update tests doc
   * tests: Add DevX traffic over CUDA DMABUF
   * tests: Add RDMA traffic over CUDA DMABUF
   * tests: Create general memory write/read functions
   * pyverbs: Support MLX5DV_UMEM_MASK_DMABUF
   * pyverbs: Generalize DmaBufMR's dmabuf input
   * providers/erdma: Add atomic operation support
   * providers/irdma: Fix RQ completion opcode
   * providers/irdma: Fix inline handling of multiple SGE's
   * tests: Fix condition of bad_flow_handling
   * tests: Fix the error message in check_mkey function
   * tests: Improve DMA memcpy bad flow test
   * tests: Skip dest port action tests if unsupported
   * tests: Adjust FDB VPort tests
   * tests: Fix excpetion handler of multithreading test
   * tests: Initialize port space with UDP on CM tests
   * tests: Move RDMACM tests to be in one direction
   * tests: Add RDMACM multicast leave test
   * debian: Remove /etc/udev/rules.d/70-persistent-ipoib.rules on upgrade
   * Fix grammar mistake of "allow doing"
   * Update library version to be 44.0
   * tests: Add a test for RDMA Write with ODP
   * tests: Add a test for RDMA Read with ODP
   * tests: Add tests for Atomic operations with ODP
   * tests: Make ODP test consistent with other traffic tests
   * tests: Import test.utils entirely for ODP test cases

+++ tests/base.py
@@ -9,6 +9,7 @@ import random
 import errno
 import stat
 import json
+import sys
 import os
 
 from pyverbs.qp import QPCap, QPInitAttrEx, QPInitAttr, QPAttr, QP
@@ -77,6 +78,7 @@ class PyverbsAPITestCase(unittest.TestCase):
         self.attr = None
         self.attr_ex = None
         self.gid_index = 0
+        self.pre_environment = {}
 
     def setUp(self):
         """
@@ -104,7 +106,23 @@ class PyverbsAPITestCase(unittest.TestCase):
     def create_context(self):
         self.ctx = d.Context(name=self.dev_name)
 
+    def set_env_variable(self, var, value):
+        """
+        Set environment variable. The current value for each variable is stored
+        and is set back at the end of the test.
+        :param var: The name of the environment variable
+        :param value: The requested new value of this environment variable
+        """
+        if var not in self.pre_environment.keys():
+            self.pre_environment[var] = os.environ.get(var)
+        os.environ[var] = value
+
     def tearDown(self):
+        for k, v in self.pre_environment.items():
+            if v is None:
+                os.environ.pop(k)
+            else:
+                os.environ[k] = v
         self.ctx.close()
 
 
@@ -146,16 +164,6 @@ class RDMATestCase(unittest.TestCase):
         self.server = None
         self.client = None
 
-    def set_env_variable(self, var, value):
-        """
-        Set environment variable. The current value for each variable is stored
-        and is set back at the end of the test.
-        :param var: The name of the environment variable
-        :param value: The requested new value of this environment variable
-        """
-        if var not in self.pre_environment.keys():
-            self.pre_environment[var] = os.environ.get(var)
-        os.environ[var] = value
 
     def is_eth_and_has_roce_hw_bug(self):
         """
@@ -325,13 +333,16 @@ class RDMACMBaseTest(RDMATestCase):
                                     .format(self.dev_name))
         is_gid_available(self.gid_index)
 
-    def two_nodes_rdmacm_traffic(self, connection_resources, test_flow,
+    def two_nodes_rdmacm_traffic(self, connection_resources, test_flow, bad_flow=False,
                                  **resource_kwargs):
         """
-        Init and manage the rdmacm test processes. If needed, terminate those
-        processes and raise an exception.
+        Init and manage the rdmacm test processes. The exit code of the
+        test processes indicates if exception was thrown.
+        {0: pass, 2: exception was thrown, 5: skip test}
+        If needed, terminate those processes and raise an exception.
         :param connection_resources: The CMConnection resources to use.
         :param test_flow: The target RDMACM flow method to run.
+        :param bad_flow: If true, traffic is expected to fail.
         :param resource_kwargs: Dict of args that specify the CMResources
                                 specific attributes. Each test case can pass
                                 here as key words the specific CMResources
@@ -353,40 +364,37 @@ class RDMACMBaseTest(RDMATestCase):
                                       'passive':False, **resource_kwargs})
         passive.start()
         active.start()
-        proc_raised_ex = False
-        for i in range(15):
-            if proc_raised_ex:
-                break
+        repeat_times=150 if not bad_flow else 3
+        proc_res = {}
+        for _ in range(repeat_times):
             for proc in [passive, active]:
-                proc.join(1)
-                if not proc.is_alive() and not self.notifier.empty():
-                    proc_raised_ex = True
-                    break
-
+                proc.join(0.1)
+                # Write the exit code of the proc.
+                if not proc.is_alive():
+                    side = 'passive' if proc == passive else 'active'
+                    if side not in proc_res.keys():
+                        proc_res[side] = proc.exitcode
         # If the processes is still alive kill them and fail the test.
         proc_killed = False
         for proc in [passive, active]:
             if proc.is_alive():
                 proc.terminate()
                 proc_killed = True
+        # Check if need to skip this test
+        for side in proc_res.keys():
+            if proc_res[side] == 5:
+                raise unittest.SkipTest(f'SkipTest occurred on {side} side')
         # Check if the test processes raise exceptions.
-        proc_res = {}
-        while not self.notifier.empty():
-            res, side = self.notifier.get()
-            proc_res[side] = res
-        for ex in proc_res.values():
-            if isinstance(ex, PyverbsRDMAError) and \
-                    ex.error_code == errno.EOPNOTSUPP:
-                        raise unittest.SkipTest(ex)
-            if isinstance(ex, unittest.case.SkipTest):
-                raise(ex)
-        if proc_res:
-            print(f'Received the following exceptions: {proc_res}')
-            if isinstance(res, Exception):
-                raise(res)
-            raise PyverbsError(res)
+        res_exception = False
+        for side in proc_res:
+            if 0 < proc_res[side] < 5:
+                res_exception = True
+        if res_exception:
+            raise Exception('Exception in active/passive side occurred')
         # Raise exeption if the test proceses was terminate.
-        if proc_killed:
+        if bad_flow and not proc_killed:
+            raise Exception('Bad flow: traffic passed which is not expected')
+        if not bad_flow and proc_killed:
             raise Exception('RDMA CM test procces is stuck, kill the test')
 
     def rdmacm_traffic(self, connection_resources=None, passive=None, **kwargs):
@@ -408,17 +416,18 @@ class RDMACMBaseTest(RDMATestCase):
             player.rdmacm_traffic()
             player.disconnect()
         except Exception as ex:
-            side = 'passive' if passive else 'active'
-            self.notifier.put((ex, side))
+            self._rdmacm_exception_handler(passive, ex)
 
     def rdmacm_multicast_traffic(self, connection_resources=None, passive=None,
-                                 extended=False, **kwargs):
+                                 extended=False, leave_test=False, **kwargs):
         """
         Run RDMACM multicast traffic between two CMIDs.
         :param connection_resources: The connection resources to use.
         :param passive: Indicate if this CMID is the passive side.
         :param extended: Use exteneded multicast join request. This request
                          allows CMID to join with specific join flags.
+        :param leave_test: Perform traffic after leaving the multicast group to
+                           ensure leave works.
         :param kwargs: Arguments to be passed to the connection_resources.
         :return: None
         """
@@ -431,9 +440,10 @@ class RDMACMBaseTest(RDMATestCase):
                                      extended=extended)
             player.rdmacm_traffic(server=passive, multicast=True)
             player.leave_multicast(mc_addr=mc_addr)
+            if leave_test:
+                player.rdmacm_traffic(server=passive, multicast=True)
         except Exception as ex:
-            side = 'passive' if passive else 'active'
-            self.notifier.put((ex, side))
+            self._rdmacm_exception_handler(passive, ex)
 
     def rdmacm_remote_traffic(self, connection_resources=None, passive=None,
                               remote_op='write', **kwargs):
@@ -455,10 +465,15 @@ class RDMACMBaseTest(RDMATestCase):
             player.remote_traffic(passive=passive, remote_op=remote_op)
             player.disconnect()
         except Exception as ex:
-            while not self.notifier.empty():
-                self.notifier.get()
-            side = 'passive' if passive else 'active'
-            self.notifier.put((ex, side))
+            self._rdmacm_exception_handler(passive, ex)
+
+    @staticmethod
+    def _rdmacm_exception_handler(passive, exception):
+        if isinstance(exception, unittest.case.SkipTest):
+            sys.exit(5)
+        side = 'passive' if passive else 'active'
+        print(f'Player {side} got: {exception}')
+        sys.exit(2)
 
 
 class BaseResources(object):
@@ -485,6 +500,13 @@ class BaseResources(object):
 
     def create_pd(self):
         self.pd = PD(self.ctx)
+
+    def mem_write(self, data, size, offset=0):
+        self.mr.write(data, size, offset)
+
+    def mem_read(self, size=None, offset=0):
+        size_ = self.msg_size if size is None else size
+        return self.mr.read(size_, offset)
 
 
 class TrafficResources(BaseResources):
